[COMMON] move ClusterBean#of(Map, MetricSensor) to test code (#1667)
