+++ common/src/test/java/org/astraea/common/cost/BrokerInputCostTest.java
@@ -22,6 +22,7 @@ import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.metrics.BeanObject;
 import org.astraea.common.metrics.MBeanClient;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.common.metrics.broker.ServerMetrics;
 import org.astraea.common.producer.Producer;
 import org.astraea.common.producer.Record;
@@ -64,7 +65,8 @@ public class BrokerInputCostTest {
     }
     var f = new BrokerInputCost();
     var clusterBean =
-        ClusterBean.of(Map.of(0, MBeanClient.of(SERVICE.jmxServiceURL())), f.metricSensor().get());
+        MetricsTestUtils.clusterBean(
+            Map.of(0, MBeanClient.of(SERVICE.jmxServiceURL())), f.metricSensor().get());
 
     Assertions.assertNotEquals(
         0, clusterBean.brokerMetrics(0, ServerMetrics.BrokerTopic.Meter.class).count());
