+++ common/src/test/java/org/astraea/common/cost/BrokerOutputCostTest.java
@@ -27,6 +27,7 @@ import org.astraea.common.consumer.Consumer;
 import org.astraea.common.consumer.ConsumerConfigs;
 import org.astraea.common.metrics.BeanObject;
 import org.astraea.common.metrics.MBeanClient;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.common.metrics.broker.ServerMetrics;
 import org.astraea.common.producer.Producer;
 import org.astraea.common.producer.Record;
@@ -79,7 +80,8 @@ public class BrokerOutputCostTest {
     }
 
     var f = new BrokerOutputCost();
-    var clusterBean = ClusterBean.of(Map.of(0, MBeanClient.local()), f.metricSensor().get());
+    var clusterBean =
+        MetricsTestUtils.clusterBean(Map.of(0, MBeanClient.local()), f.metricSensor().get());
 
     Assertions.assertNotEquals(
         0, clusterBean.brokerMetrics(0, ServerMetrics.BrokerTopic.Meter.class).count());
