+++ common/src/test/java/org/astraea/common/cost/MemoryCostTest.java
@@ -23,6 +23,7 @@ import java.util.Objects;
 import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.metrics.MBeanClient;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.common.metrics.platform.HasJvmMemory;
 import org.astraea.common.metrics.platform.JvmMemory;
 import org.junit.jupiter.api.Assertions;
@@ -52,7 +53,8 @@ public class MemoryCostTest {
   @Test
   void testSensor() {
     var f = new MemoryCost();
-    var clusterBean = ClusterBean.of(Map.of(0, MBeanClient.local()), f.metricSensor().get());
+    var clusterBean =
+        MetricsTestUtils.clusterBean(Map.of(0, MBeanClient.local()), f.metricSensor().get());
     Assertions.assertTrue(clusterBean.brokerMetrics(0, JvmMemory.class).allMatch(Objects::nonNull));
 
     // Test if we can get "used memory" and "max memory".
