+++ common/src/test/java/org/astraea/common/metrics/broker/ControllerMetricsTest.java
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.util.Arrays;
 import java.util.Locale;
 import org.astraea.common.metrics.MBeanClient;
-import org.astraea.common.metrics.MetricsTestUtil;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.it.Service;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -48,7 +48,7 @@ class ControllerMetricsTest {
   @EnumSource(ControllerMetrics.Controller.class)
   void testController(ControllerMetrics.Controller controller) {
     var gauge = controller.fetch(MBeanClient.local());
-    MetricsTestUtil.validate(gauge);
+    MetricsTestUtils.validate(gauge);
     Assertions.assertEquals(controller, gauge.type());
   }
