+++ common/src/test/java/org/astraea/common/metrics/broker/NetworkMetricsTest.java
@@ -17,7 +17,7 @@
 package org.astraea.common.metrics.broker;
 
 import org.astraea.common.metrics.MBeanClient;
-import org.astraea.common.metrics.MetricsTestUtil;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.it.Service;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -45,13 +45,13 @@ public class NetworkMetricsTest {
   @EnumSource(value = NetworkMetrics.Request.class)
   void testRequestTotalTimeMs(NetworkMetrics.Request request) {
     var histogram = request.fetch(MBeanClient.local());
-    MetricsTestUtil.validate(histogram);
+    MetricsTestUtils.validate(histogram);
   }
 
   @Test
   void testAllEnumNameUnique() {
     Assertions.assertTrue(
-        MetricsTestUtil.metricDistinct(
+        MetricsTestUtils.metricDistinct(
             NetworkMetrics.Request.values(), NetworkMetrics.Request::metricName));
   }
 }
