+++ common/src/test/java/org/astraea/common/metrics/broker/ReplicaManagerMetricsTest.java
@@ -17,7 +17,7 @@
 package org.astraea.common.metrics.broker;
 
 import org.astraea.common.metrics.MBeanClient;
-import org.astraea.common.metrics.MetricsTestUtil;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.it.Service;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -46,14 +46,14 @@ public class ReplicaManagerMetricsTest {
   void testBrokerTopic(ServerMetrics.ReplicaManager rm) {
     var gauge = rm.fetch(MBeanClient.local());
     Assertions.assertEquals(0, gauge.value());
-    MetricsTestUtil.validate(gauge);
+    MetricsTestUtils.validate(gauge);
     Assertions.assertEquals(rm, gauge.type());
   }
 
   @Test
   void testAllEnumNameUnique() {
     Assertions.assertTrue(
-        MetricsTestUtil.metricDistinct(
+        MetricsTestUtils.metricDistinct(
             ServerMetrics.ReplicaManager.values(), ServerMetrics.ReplicaManager::metricName));
   }
 }
