+++ common/src/test/java/org/astraea/common/metrics/broker/ServerTopicMetricsTest.java
@@ -17,7 +17,7 @@
 package org.astraea.common.metrics.broker;
 
 import org.astraea.common.metrics.MBeanClient;
-import org.astraea.common.metrics.MetricsTestUtil;
+import org.astraea.common.metrics.MetricsTestUtils;
 import org.astraea.it.Service;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -44,6 +44,6 @@ public class ServerTopicMetricsTest {
   @EnumSource(value = ServerMetrics.BrokerTopic.class)
   void testRequestBrokerTopicMetrics(ServerMetrics.BrokerTopic metric) {
     var meter = metric.fetch(MBeanClient.local());
-    MetricsTestUtil.validate(meter);
+    MetricsTestUtils.validate(meter);
   }
 }
